requirements.txt
discord.py==2.6.4
requests==2.32.5
python-dotenv==1.2.1

# need natural language processing
nlp
spacy
